// src/components/works/types.ts
export interface Props {
  myWorks: string;
  hardSkillsTitle: string;
  experienceTitle: string;
  achievementsTitle: string;
  worksContent1: string;
  worksContent2: string;
  worksContent3: string;
  projects: string;
  worksContent4: string;
  worksContent5: string;
  worksListLKS: string;
  worksItem1: string;
  worksItem2: string;
  worksItem3: string;
  worksItem4: string;
  worksItem5: string;
  worksListTools: string;
  worksItem6: string;
  worksItem7: string;
  worksItem8: string;
  worksContent6: string;
  worksContent7: string;
  workstools: string;
  frontendDev: string;
  tools: string;
  gitHub: string;
  tools2: string;
  worksItem10: string;
  javaScript: string;
  typeScript: string;
  react: string;
  next: string;
  rtk: string;
  swr: string;
  worksItem9: string;
  bem: string;
  css: string;
  tools3: string;
}